backend/bmp_service/scripts/parse_stopp_pdf.py
"""Extract STOPP/START v3 rules from PDF and dump to YAML.

This lightweight CLI is designed for one-shot generation of the YAML file
`data/stopp_start_v3.yaml` used by the Hygie-AI rules engine.  It purposely
implements **simple heuristics** rather than perfect parsing: the goal is to
obtain a *good-enough* structured representation that can be manually
reviewed.

The implementation follows Hygie-AI dev rules:
• Each function ≤50 lines
• ≥2 assertions per function
• Minimal heap allocations, explicit termination conditions.
"""

from __future__ import annotations

import argparse
import re
import sys
from pathlib import Path
from typing import Dict, List

import yaml  # type: ignore
from pdfminer.high_level import extract_text  # type: ignore

# ---------------------------------------------------------------------------
# Constants
# ---------------------------------------------------------------------------

# Ignore common French stop-words and generic terms that are **not** drug names
_STOPWORDS: set[str] = {
    "DE",
    "DES",
    "DU",
    "LA",
    "LE",
    "LES",
    "ET",
    "UN",
    "UNE",
    "CHEZ",
    "PATIENT",
    "SI",
    "AVEC",
    "SANS",
    "CHRONIQUE",
    "MOIS",
    "JOUR",
    "PLUS",
    "ANS",
}

# Match rule IDs that may appear as "A1.", "A 1", "A1)" etc.
_RULE_RE = re.compile(r"^\s*([A-Q])\s?([0-9]{1,2})[)\.\-\s]+(?P<desc>.+)")
_AGE_RE = re.compile(r"(\d{2})\s*(?:ans|an)", re.I)
_DCI_RE = re.compile(r"[A-ZÀÂÉÈÊÎÔÙÛÇ][A-ZÀÂÉÈÊÎÔÙÛÇ\-]{2,}")  # ≥3 letters

# ---------------------------------------------------------------------------
# Core parsing helpers – Keep each ≤50 lines
# ---------------------------------------------------------------------------


def _extract_rules(text: str) -> List[Dict[str, object]]:  # noqa: D401
    """Return list of structured rules extracted from *text*."""
    assert text, "PDF text is empty"
    lines = [ln.strip() for ln in text.splitlines() if ln.strip()]
    assert lines, "No non-empty lines after split"
    # Isolate STOPP criteria section by finding second occurrence of TOC title
    start_idx = 0
    stopp_count = 0
    for i, ln in enumerate(lines):
        if re.match(r"^Liste des critères\s+STOPP", ln, flags=re.I):
            stopp_count += 1
            if stopp_count == 2:
                start_idx = i + 1
                break
    # fallback to first occurrence
    if start_idx == 0:
        for i, ln in enumerate(lines):
            if re.match(r"^Liste des critères\s+STOPP", ln, flags=re.I):
                start_idx = i + 1
                break
    # Find START section to mark end of STOPP
    end_idx = len(lines)
    start_count = 0
    for i, ln in enumerate(lines[start_idx:], start_idx):
        if re.match(r"^Liste des critères\s+START", ln, flags=re.I):
            start_count += 1
            if start_count == 2:
                end_idx = i
                break
    # fallback to first occurrence
    if end_idx == len(lines):
        for i, ln in enumerate(lines[start_idx:], start_idx):
            if re.match(r"^Liste des critères\s+START", ln, flags=re.I):
                end_idx = i
                break
    lines = lines[start_idx:end_idx]
    assert lines, "No STOPP criteria section found"

    results: List[Dict[str, object]] = []
    for idx, line in enumerate(lines):
        match = _RULE_RE.match(line)
        if not match:
            continue
        rid = match.group(1) + match.group(2)
        desc = match.group("desc").strip()

        # If description is very short, try to concatenate next line(s)
        if len(desc) < 20 and idx + 1 < len(lines):
            nxt = lines[idx + 1].strip()
            if nxt and not _RULE_RE.match(nxt):
                desc += " " + nxt

        # Substances: heuristics – all uppercase tokens ≥3 letters not in stopwords
        subs_raw = {tok.upper() for tok in _DCI_RE.findall(desc)}
        subs = sorted(s for s in subs_raw if s not in _STOPWORDS)

        # Age criteria
        age_min = None
        age_m = _AGE_RE.search(desc)
        if age_m:
            age_min = int(age_m.group(1))

        rule: Dict[str, object] = {
            "id": rid,
            "description": desc,
            "substances": subs,
            "enabled": True,
        }
        if age_min is not None:
            rule["age_min"] = age_min

        results.append(rule)

        # Explicit upper bound on iterations for dev rule #2 (but while loop not used)
        if len(results) > 500:  # safety – STOPP+START ≤350 expected
            break

    # Fallback parsing if too few rules extracted
    if len(results) < 10:
        flat = " ".join(text.split())
        # Multiline fallback: capture entire criterion text until the next ID or end of doc
        F_RE = re.compile(
            r"\b([A-Q])\s?([0-9]{1,2})[)\.\s]+(.*?)(?=\b[A-Q]\s?[0-9]{1,2}[)\.\s]|$)",
            re.S,
        )
        fallback = []
        for m in F_RE.finditer(flat):
            rid = f"{m.group(1)}{m.group(2)}"
            desc = " ".join(m.group(3).split())
            subs_raw = {tok.upper() for tok in _DCI_RE.findall(desc)}
            subs = sorted(s for s in subs_raw if s not in _STOPWORDS)
            age_m = _AGE_RE.search(desc)
            rule = {"id": rid, "description": desc, "substances": subs, "enabled": True}
            if age_m:
                rule["age_min"] = int(age_m.group(1))
            fallback.append(rule)
        if fallback:
            results = fallback
    return results


# -- Plain-text STOPP parser --------------------------------------------------
def _extract_rules_plain(text: str) -> List[Dict[str, object]]:  # noqa: D401
    """Parse raw *text* where each section is introduced by
    "Section X:" followed by numbered bullets. Returns list of rules.
    Obeys Hygie-AI dev rules: ≤50 lines, ≥2 assertions.
    """
    assert text, "Input text empty"
    lines = [ln.strip() for ln in text.splitlines()]
    assert lines, "No lines after split"

    results: List[Dict[str, object]] = []
    letter: str | None = None
    num: int | None = None
    parts: List[str] = []
    # Track whether rule belongs to STOPP or START phase
    phase: str = "STOPP"

    def _flush() -> None:
        nonlocal num, parts
        if letter and num is not None and parts:
            desc = " ".join(parts).strip()
            rid = f"{letter}{num}"
            subs_raw = {t.upper() for t in _DCI_RE.findall(desc)}
            subs = sorted(s for s in subs_raw if s not in _STOPWORDS)
            rule: Dict[str, object] = {
                "id": rid,
                "description": desc,
                "substances": subs,
                "enabled": True,
                "phase": phase,
            }
            age_m = _AGE_RE.search(desc)
            if age_m:
                rule["age_min"] = int(age_m.group(1))
            results.append(rule)
        parts = []

    sec_re = re.compile(r"^Section\s+([A-Q])[:\s]", re.I)
    bullet_re = re.compile(r"^(\d{1,3})\.\s*(.*)")

    for ln in lines:
        # Switch to START phase when encountering START header
        if re.match(r"^Screening Tool to Alert to Right Treatment", ln, flags=re.I):
            _flush()
            phase = "START"
            letter = None
            num = None
            continue
        if not ln:
            continue
        m_sec = sec_re.match(ln)
        if m_sec:
            _flush()
            letter = m_sec.group(1).upper()
            num = None
            continue
        m_bul = bullet_re.match(ln)
        if m_bul and letter:
            _flush()
            num = int(m_bul.group(1))
            first = m_bul.group(2).strip()
            parts = [first] if first else []
        elif num is not None:
            parts.append(ln)

    _flush()
    return results


def _parse_pdf(pdf_path: Path) -> List[Dict[str, object]]:  # noqa: D401
    """Parse PDF or plain-text file containing STOPP criteria."""
    assert pdf_path.exists(), f"File not found: {pdf_path}"
    if pdf_path.suffix.lower() == ".pdf":
        text = extract_text(str(pdf_path))
        return _extract_rules(text)
    # treat everything else as raw text
    text = pdf_path.read_text(encoding="utf-8", errors="ignore")
    return _extract_rules_plain(text)


# ---------------------------------------------------------------------------
# CLI entry-point
# ---------------------------------------------------------------------------


def _cli(argv: List[str]) -> None:  # noqa: D401
    parser = argparse.ArgumentParser(
        description="Parse STOPP/START PDF or plain-text → YAML"
    )
    parser.add_argument(
        "pdf", type=Path, help="Path to carnet STOPP/START PDF or plain-text file"
    )
    parser.add_argument(
        "-o",
        "--output",
        type=Path,
        default=Path(__file__).resolve().parents[1] / "data" / "stopp_start_v3.yaml",
        help="Output YAML file (default: data/stopp_start_v3.yaml)",
    )
    args = parser.parse_args(argv)

    rules = _parse_pdf(args.pdf)
    assert rules, "No rules extracted – check PDF layout"

    # Write YAML – minimal heap allocations via generator
    with args.output.open("w", encoding="utf-8") as fh:
        yaml.dump(rules, fh, allow_unicode=True, sort_keys=False)
    print(f"✔️  {len(rules)} rules written to {args.output}")


# ---------------------------------------------------------------------------
# Module guard
# ---------------------------------------------------------------------------

if __name__ == "__main__":  # pragma: no cover
    _cli(sys.argv[1:])
